fix: test

+++ server/models/mock.js
@@ -3,17 +3,17 @@
 module.exports = function(sequelize, DataTypes) {
   return sequelize.define('mock', {
     'id': {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.INTEGER(10).UNSIGNED,
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
     },
     'project_id': {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.INTEGER(10).UNSIGNED,
       allowNull: true
     },
     'uid': {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.INTEGER(10).UNSIGNED,
       allowNull: true,
       unique: true
     },
@@ -23,12 +23,12 @@ module.exports = function(sequelize, DataTypes) {
       defaultValue: ''
     },
     'method': {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.INTEGER(10).UNSIGNED,
       allowNull: true,
       defaultValue: '1'
     },
     'delay': {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.INTEGER(10).UNSIGNED,
       allowNull: true,
       defaultValue: '0'
     },
