+++ server/models/user_project.js
@@ -3,17 +3,17 @@
 module.exports = function(sequelize, DataTypes) {
   return sequelize.define('user_project', {
     'id': {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.INTEGER(10).UNSIGNED,
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
     },
     'uid': {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: true
     },
     'project_id': {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: true
     },
     'created_at': {
